rotation: remove orphan rotations
An orphan rotation is found when a jump pole produces a change in the
fixed plate, but there are no more references to that plate, for
example:

	505 0.0 0.0 0.0 0.0 501
	505 50.0 -28.83 -123.27 40.16 501
	505 65.0 -33.6 -123.6 75.56 501
	505 65.0 -17.21 -138.31 116.59 000
	505 65.0 -22.55 -127.64 106.34 503
	505 96.0 -22.55 -127.64 106.34 503

In 65.0 Ma, there are three jumps, but one, the fixed plate '000' is
orphan, so that jump will be deleted.

This scenario is quite infrequent but happens in the PaleoMap rotation
file. Instead of asking the user to remove it, the package will remove
it automatically.